Move CourseWidget to academics admin
